refactor(server): export tests to files
